Simplify auth logic
